Fixed Player Selection
I fixed the issue of having the ref choose two players. The
first problem I solved by using .includes which allowed me
to tell if the user was actually in the system. The other issue
was making sure that the ref also did not try and put in the
same name for both players.